requirements.txt
Flask==2.3.3
line-bot-sdk==2.4.1
gunicorn==20.0.4
google-auth==1.25.0
google-cloud-vision==2.1.0
matplotlib==3.7.5
python-dotenv==0.19.0
japanize_matplotlib==1.1.3
flask-jwt-extended==4.4.4
Authlib==1.2.0
WTForms==3.0.1
flask_bcrypt==1.0.1
PyMongo==4.3.3
Pillow==10.2.0
certifi==2024.2.2